+++ ll/src/blocking/vdm.rs
@@ -13,6 +13,8 @@ use crate::spi::{vdm_header, AccessMode};
 use embedded_hal::digital::v2::OutputPin;
 
 /// W5500 blocking variable data length implementation.
+#[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct W5500<SPI, CS> {
     /// SPI bus.
     spi: SPI,
@@ -22,6 +24,7 @@ pub struct W5500<SPI, CS> {
 
 /// W5500 blocking implementation error type.
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Error<SpiError, PinError> {
     /// SPI bus error wrapper.
     Spi(SpiError),