+++ ll/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `defmt::Display` for all enumerations.
+- Added `defmt::Display` for `fdm::W5500`, `vdm::W5500`, and `vdm::Error`.
+
 ## [0.8.0] - 2021-05-08
 ### Added
 - Added `Display` for `SocketInterruptMask`.